refactor(nuxi): move packages functions to `utils`
